split geo-related code to separate files
