+++ src/vector/geometry.rs
@@ -4,7 +4,6 @@ use std::ffi::CString;
 use std::cell::RefCell;
 use utils::_string;
 use vector::ogr;
-use geo;
 
 /// OGR Geometry
 pub struct Geometry {
@@ -126,7 +125,7 @@ impl Geometry {
         return unsafe { Geometry::with_c_geometry(c_geom, true) };
     }
 
-    unsafe fn _get_geometry(&self, n: usize) -> Geometry {
+    pub unsafe fn _get_geometry(&self, n: usize) -> Geometry {
         // get the n-th sub-geometry as a non-owned Geometry; don't keep this
         // object for long.
         let c_geom = ogr::OGR_G_GetGeometryRef(self.c_geometry(), n as c_int);
@@ -144,322 +143,11 @@ impl Geometry {
     }
 }
 
-impl geo::ToGeo for Geometry {
-    fn to_geo(&self) -> geo::Geometry {
-        let geometry_type = unsafe { ogr::OGR_G_GetGeometryType(self.c_geometry()) };
-
-        let ring = |n: usize| {
-            let ring = unsafe { self._get_geometry(n) };
-            return match ring.to_geo() {
-                geo::Geometry::LineString(r) => r,
-                _ => panic!("Expected to get a LineString")
-            };
-        };
-
-        match geometry_type {
-            ogr::WKB_POINT => {
-                let (x, y, _) = self.get_point(0);
-                geo::Geometry::Point(geo::Point(geo::Coordinate{x: x, y: y}))
-            },
-            ogr::WKB_MULTIPOINT => {
-                let point_count = unsafe { ogr::OGR_G_GetGeometryCount(self.c_geometry()) } as usize;
-                let coords = (0..point_count)
-                    .map(|n| {
-                        match unsafe { self._get_geometry(n) }.to_geo() {
-                            geo::Geometry::Point(p) => p,
-                            _ => panic!("Expected to get a Point")
-                        }
-                    })
-                    .collect();
-                geo::Geometry::MultiPoint(geo::MultiPoint(coords))
-            },
-            ogr::WKB_LINESTRING => {
-                let coords = self.get_point_vec().iter()
-                    .map(|&(x, y, _)| geo::Point(geo::Coordinate{x: x, y: y}))
-                    .collect();
-                geo::Geometry::LineString(geo::LineString(coords))
-            },
-            ogr::WKB_MULTILINESTRING => {
-                let string_count = unsafe { ogr::OGR_G_GetGeometryCount(self.c_geometry()) } as usize;
-                let strings = (0..string_count)
-                    .map(|n| {
-                        match unsafe { self._get_geometry(n) }.to_geo() {
-                            geo::Geometry::LineString(s) => s,
-                            _ => panic!("Expected to get a LineString")
-                        }
-                    })
-                    .collect();
-                geo::Geometry::MultiLineString(geo::MultiLineString(strings))
-            },
-            ogr::WKB_POLYGON => {
-                let ring_count = unsafe { ogr::OGR_G_GetGeometryCount(self.c_geometry()) } as usize;
-                let outer = ring(0);
-                let holes = (1..ring_count).map(|n| ring(n)).collect();
-                geo::Geometry::Polygon(geo::Polygon(outer, holes))
-            },
-            ogr::WKB_MULTIPOLYGON => {
-                let string_count = unsafe { ogr::OGR_G_GetGeometryCount(self.c_geometry()) } as usize;
-                let strings = (0..string_count)
-                    .map(|n| {
-                        match unsafe { self._get_geometry(n) }.to_geo() {
-                            geo::Geometry::Polygon(s) => s,
-                            _ => panic!("Expected to get a Polygon")
-                        }
-                    })
-                    .collect();
-                geo::Geometry::MultiPolygon(geo::MultiPolygon(strings))
-            },
-            ogr::WKB_GEOMETRYCOLLECTION => {
-                let item_count = unsafe { ogr::OGR_G_GetGeometryCount(self.c_geometry()) } as usize;
-                let geometry_list = (0..item_count)
-                    .map(|n| unsafe { self._get_geometry(n) }.to_geo())
-                    .collect();
-                geo::Geometry::GeometryCollection(geo::GeometryCollection(geometry_list))
-            }
-            _ => panic!("Unknown geometry type")
-        }
-    }
-}
-
 impl Drop for Geometry {
     fn drop(&mut self) {
         if self.owned {
             let c_geometry = self.c_geometry_ref.borrow();
             unsafe { ogr::OGR_G_DestroyGeometry(c_geometry.unwrap() as *mut ()) };
         }
-    }
-}
-
-
-pub trait ToGdal {
-    fn to_gdal(&self) -> Geometry;
-}
-
-
-impl ToGdal for geo::Point {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_POINT);
-        let &geo::Point(coordinate) = self;
-        geom.set_point_2d(0, (coordinate.x, coordinate.y));
-        return geom;
-    }
-}
-
-impl ToGdal for geo::MultiPoint {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_MULTIPOINT);
-        let &geo::MultiPoint(ref point_list) = self;
-        for point in point_list.iter() {
-            geom.add_geometry(point.to_gdal());
-        }
-        return geom;
-    }
-}
-
-fn geometry_with_points(wkb_type: c_int, points: &geo::LineString) -> Geometry {
-    let mut geom = Geometry::empty(wkb_type);
-    let &geo::LineString(ref linestring) = points;
-    for (i, &geo::Point(coordinate)) in linestring.iter().enumerate() {
-        geom.set_point_2d(i, (coordinate.x, coordinate.y));
-    }
-    return geom;
-}
-
-impl ToGdal for geo::LineString {
-    fn to_gdal(&self) -> Geometry {
-        geometry_with_points(ogr::WKB_LINESTRING, self)
-    }
-}
-
-impl ToGdal for geo::MultiLineString {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_MULTILINESTRING);
-        let &geo::MultiLineString(ref point_list) = self;
-        for point in point_list.iter() {
-            geom.add_geometry(point.to_gdal());
-        }
-        return geom;
-    }
-}
-
-impl ToGdal for geo::Polygon {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_POLYGON);
-        let &geo::Polygon(ref outer, ref holes) = self;
-        geom.add_geometry(geometry_with_points(ogr::WKB_LINEARRING, outer));
-        for ring in holes.iter() {
-            geom.add_geometry(geometry_with_points(ogr::WKB_LINEARRING, ring));
-        }
-        return geom;
-    }
-}
-
-impl ToGdal for geo::MultiPolygon {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_MULTIPOLYGON);
-        let &geo::MultiPolygon(ref polygon_list) = self;
-        for polygon in polygon_list.iter() {
-            geom.add_geometry(polygon.to_gdal());
-        }
-        return geom;
-    }
-}
-
-impl ToGdal for geo::GeometryCollection {
-    fn to_gdal(&self) -> Geometry {
-        let mut geom = Geometry::empty(ogr::WKB_GEOMETRYCOLLECTION);
-        let &geo::GeometryCollection(ref item_list) = self;
-        for item in item_list.iter() {
-            geom.add_geometry(item.to_gdal());
-        }
-        return geom;
-    }
-}
-
-impl ToGdal for geo::Geometry {
-    fn to_gdal(&self) -> Geometry {
-        return match *self {
-            geo::Geometry::Point(ref c) => c.to_gdal(),
-            geo::Geometry::MultiPoint(ref c) => c.to_gdal(),
-            geo::Geometry::LineString(ref c) => c.to_gdal(),
-            geo::Geometry::MultiLineString(ref c) => c.to_gdal(),
-            geo::Geometry::Polygon(ref c) => c.to_gdal(),
-            geo::Geometry::MultiPolygon(ref c) => c.to_gdal(),
-            geo::Geometry::GeometryCollection(ref c) => c.to_gdal(),
-        }
-    }
-}
-
-
-#[cfg(test)]
-mod tests {
-    use vector::{Geometry, ToGdal};
-    use geo;
-    use geo::ToGeo;
-
-    #[test]
-    fn test_import_export_point() {
-        let wkt = "POINT (1 2)";
-        let coord = geo::Coordinate{x: 1., y: 2.};
-        let geo = geo::Geometry::Point(geo::Point(coord));
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    #[test]
-    fn test_import_export_multipoint() {
-        let wkt = "MULTIPOINT (0 0,0 1,1 2)";
-        let coord = vec!(
-            geo::Point(geo::Coordinate{x: 0., y: 0.}),
-            geo::Point(geo::Coordinate{x: 0., y: 1.}),
-            geo::Point(geo::Coordinate{x: 1., y: 2.}),
-        );
-        let geo = geo::Geometry::MultiPoint(geo::MultiPoint(coord));
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    #[test]
-    fn test_import_export_linestring() {
-        let wkt = "LINESTRING (0 0,0 1,1 2)";
-        let coord = vec!(
-            geo::Point(geo::Coordinate{x: 0., y: 0.}),
-            geo::Point(geo::Coordinate{x: 0., y: 1.}),
-            geo::Point(geo::Coordinate{x: 1., y: 2.}),
-        );
-        let geo = geo::Geometry::LineString(geo::LineString(coord));
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    #[test]
-    fn test_import_export_multilinestring() {
-        let wkt = "MULTILINESTRING ((0 0,0 1,1 2),(3 3,3 4,4 5))";
-        let strings = vec!(
-            geo::LineString(vec!(
-                geo::Point(geo::Coordinate{x: 0., y: 0.}),
-                geo::Point(geo::Coordinate{x: 0., y: 1.}),
-                geo::Point(geo::Coordinate{x: 1., y: 2.}),
-            )),
-            geo::LineString(vec!(
-                geo::Point(geo::Coordinate{x: 3., y: 3.}),
-                geo::Point(geo::Coordinate{x: 3., y: 4.}),
-                geo::Point(geo::Coordinate{x: 4., y: 5.}),
-            )),
-        );
-        let geo = geo::Geometry::MultiLineString(geo::MultiLineString(strings));
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    fn square(x0: isize, y0: isize, x1: isize, y1: isize) -> geo::LineString {
-        geo::LineString(vec!(
-            geo::Point(geo::Coordinate{x: x0 as f64, y: y0 as f64}),
-            geo::Point(geo::Coordinate{x: x0 as f64, y: y1 as f64}),
-            geo::Point(geo::Coordinate{x: x1 as f64, y: y1 as f64}),
-            geo::Point(geo::Coordinate{x: x1 as f64, y: y0 as f64}),
-            geo::Point(geo::Coordinate{x: x0 as f64, y: y0 as f64}),
-        ))
-    }
-
-    #[test]
-    fn test_import_export_polygon() {
-        let wkt = "POLYGON ((0 0,0 5,5 5,5 0,0 0),\
-                            (1 1,1 2,2 2,2 1,1 1),\
-                            (3 3,3 4,4 4,4 3,3 3))";
-        let outer = square(0, 0, 5, 5);
-        let holes = vec!(square(1, 1, 2, 2), square(3, 3, 4, 4));
-        let geo = geo::Geometry::Polygon(geo::Polygon(outer, holes));
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    #[test]
-    fn test_import_export_multipolygon() {
-        let wkt = "MULTIPOLYGON (\
-            ((0 0,0 5,5 5,5 0,0 0),\
-             (1 1,1 2,2 2,2 1,1 1),\
-             (3 3,3 4,4 4,4 3,3 3)),\
-            ((4 4,4 9,9 9,9 4,4 4),\
-             (5 5,5 6,6 6,6 5,5 5),\
-             (7 7,7 8,8 8,8 7,7 7))\
-            )";
-        let multipolygon = geo::MultiPolygon(vec!(
-            geo::Polygon(
-                square(0, 0, 5, 5),
-                vec!(square(1, 1, 2, 2), square(3, 3, 4, 4)),
-            ),
-            geo::Polygon(
-                square(4, 4, 9, 9),
-                vec!(square(5, 5, 6, 6), square(7, 7, 8, 8)),
-            ),
-        ));
-        let geo = geo::Geometry::MultiPolygon(multipolygon);
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
-    }
-
-    #[test]
-    fn test_import_export_geometrycollection() {
-        let wkt = "GEOMETRYCOLLECTION (POINT (1 2),LINESTRING (0 0,0 1,1 2))";
-        let coord = geo::Coordinate{x: 1., y: 2.};
-        let point = geo::Geometry::Point(geo::Point(coord));
-        let coords = vec!(
-            geo::Point(geo::Coordinate{x: 0., y: 0.}),
-            geo::Point(geo::Coordinate{x: 0., y: 1.}),
-            geo::Point(geo::Coordinate{x: 1., y: 2.}),
-        );
-        let linestring = geo::Geometry::LineString(geo::LineString(coords));
-        let collection = geo::GeometryCollection(vec!(point, linestring));
-        let geo = geo::Geometry::GeometryCollection(collection);
-
-        assert_eq!(Geometry::from_wkt(wkt).to_geo(), geo);
-        assert_eq!(geo.to_gdal().wkt(), wkt);
     }
 }
